requirements.txt
streamlit
phi
yfinance
duckduckgo-search
python-dotenv